FaceAnalysisEngine works (also exmaple faces and the threshold need to be tuned)

+++ projects/Core/src/com/gpigc/core/analysis/engine/FaceAnalysisEngine.java
@@ -3,7 +3,6 @@
  */
 package com.gpigc.core.analysis.engine;
 
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -129,7 +128,7 @@ public class FaceAnalysisEngine extends AnalysisEngine {
 	public static Mat parseFace(String faceMatrixStr) {
 		// Parse face matrix
 		String[] elements = faceMatrixStr.split(",");
-		Mat faceMatrix = new Mat(new Size(elements.length, 1), CvType.CV_64FC1);
+		Mat faceMatrix = new Mat(new Size(elements.length, 1), CvType.CV_32FC1);
 		for (int i = 0; i < elements.length; i++) {
 			double elementValue = Double.parseDouble(elements[i]);
 			faceMatrix.put(0, i, elementValue);
